added simple factory pattern for the four charts
All four charts 'work' now and the factory pattern
made it easier to organize them in code. There's
still some stuff I think might be code duplication
but the four components are pretty independent.
The 'All Games Chart' is currently the most suspect
and the one I want to detail the most.