add percentage labels to the win-loss pie chart
Doing this without the chartjs plugin for datalabels was a
losing battle - many of the StackOverflow answers were
focused on JS implementations and relied on the 'meta'
property which appears to be inaccessible. Thankfully the
plugin is also written by the Chartjs dev team, so the
integration is quite seamless. Better to rely on a library and
have a solution than insist on rolling your own and
learning nothing from doing so.