initial tooltip customization for the line charts
This is a pretty big step forward for rendering the
metadata for each point in both line charts. I want to
see if I can add some flare/pizazz with character icons
in the tooltip. I might need to commission someone to
make these icons, but all the cool data that people would
want to see is there. Also fixed up some content for the
about page markdown and changed 'Bison' to 'M. Bison'
for naming consistency