ensure that if users resize window and restore to original state, chart content adjusts consistently in both resize operations. begin work on more detailed win/loss chart.
ChartJS has noted on their website that in order for
the responsive part of their chart library to work
consistently, some custom CSS must be applied to a
div that wraps the chart canvas. The responsiveness
only triggers both ways when the chart canvas is wrapped
in a parent element, as ChartJS sources width and height
values from its parent container when a resize event occurs.

Fixed an issue where if a user zoomed in on
the very first game of all games, the previousDatum var
would be null as there's no prior data to source.

Started fleshing out the codebehind for the winlosschart
component to allow users to see my total win/loss and
my win/loss per character.